refactor: switch to pgsql, add session model

+++ models/enter.go
@@ -5,7 +5,7 @@ package models
 import "time"
 
 type Model struct {
-	ID        uint      `gorm:"primaryKey" json:"id"`
+	ID        int64     `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time `json:"createdAt"`
 	UpdatedAt time.Time `json:"updatedAt"`
 }
